Created util method in Parsing#toAbstractSyntaxTree
