+++ src/main/java/com/cinchapi/ccl/ConcourseParser.java
@@ -15,8 +15,6 @@
  */
 package com.cinchapi.ccl;
 
-import java.util.ArrayDeque;
-import java.util.Deque;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Queue;
@@ -28,7 +26,6 @@ import javax.annotation.concurrent.ThreadSafe;
 import org.apache.commons.lang.StringUtils;
 
 import com.cinchapi.ccl.grammar.ConjunctionSymbol;
-import com.cinchapi.ccl.grammar.Expression;
 import com.cinchapi.ccl.grammar.KeySymbol;
 import com.cinchapi.ccl.grammar.OperatorSymbol;
 import com.cinchapi.ccl.grammar.ParenthesisSymbol;
@@ -37,9 +34,6 @@ import com.cinchapi.ccl.grammar.Symbol;
 import com.cinchapi.ccl.grammar.TimestampSymbol;
 import com.cinchapi.ccl.grammar.ValueSymbol;
 import com.cinchapi.ccl.syntax.AbstractSyntaxTree;
-import com.cinchapi.ccl.syntax.AndTree;
-import com.cinchapi.ccl.syntax.ExpressionTree;
-import com.cinchapi.ccl.syntax.OrTree;
 import com.cinchapi.ccl.type.Operator;
 import com.cinchapi.ccl.util.NaturalLanguage;
 import com.cinchapi.common.base.AnyStrings;
@@ -94,54 +88,7 @@ final class ConcourseParser extends Parser {
 
     @Override
     public AbstractSyntaxTree parse() {
-        List<Symbol> symbols = tokenize();
-        Deque<Symbol> operatorStack = new ArrayDeque<Symbol>();
-        Deque<AbstractSyntaxTree> operandStack = new ArrayDeque<AbstractSyntaxTree>();
-        symbols = Parsing.groupExpressions(symbols);
-        main: for (Symbol symbol : symbols) {
-            if(symbol == ParenthesisSymbol.LEFT) {
-                operatorStack.push(symbol);
-            }
-            else if(symbol == ParenthesisSymbol.RIGHT) {
-                while (!operatorStack.isEmpty()) {
-                    Symbol popped = operatorStack.pop();
-                    if(popped == ParenthesisSymbol.LEFT) {
-                        continue main;
-                    }
-                    else {
-                        addAbstractSyntaxTreeNode(operandStack, popped);
-                    }
-                }
-                throw new SyntaxException(AnyStrings.format(
-                        "Syntax error in {}: Mismatched parenthesis", symbols));
-            }
-            else if(symbol instanceof ConjunctionSymbol) {
-                final ConjunctionSymbol con1 = (ConjunctionSymbol) symbol;
-                Symbol symbol2;
-                while (!operatorStack.isEmpty()
-                        && (symbol2 = operatorStack.peek()) != null
-                        && symbol2 instanceof ConjunctionSymbol) {
-                    ConjunctionSymbol con2 = (ConjunctionSymbol) symbol2;
-                    if((!con1.isRightAssociative()
-                            && con1.comparePrecedence(con2) == 0)
-                            || con1.comparePrecedence(con2) < 0) {
-                        operatorStack.pop();
-                        addAbstractSyntaxTreeNode(operandStack, con2);
-                    }
-                    else {
-                        break;
-                    }
-                }
-                operatorStack.push(symbol);
-            }
-            else if(symbol instanceof Expression) {
-                operandStack.push(new ExpressionTree((Expression) symbol));
-            }
-        }
-        while (!operatorStack.isEmpty()) {
-            addAbstractSyntaxTreeNode(operandStack, operatorStack.pop());
-        }
-        return operandStack.pop();
+        return Parsing.toAbstractSyntaxTree(tokenize());
     }
 
     @Override
@@ -251,26 +198,6 @@ final class ConcourseParser extends Parser {
     @Override
     public Object transformValue(String token) {
         return valueTransformFunction.apply(token);
-    }
-
-    /**
-     * An the appropriate {@link AbstractSyntaxTree} node to the {@code stack}
-     * based on
-     * {@code operator}.
-     * 
-     * @param stack
-     * @param operator
-     */
-    private void addAbstractSyntaxTreeNode(Deque<AbstractSyntaxTree> stack,
-            Symbol operator) {
-        AbstractSyntaxTree right = stack.pop();
-        AbstractSyntaxTree left = stack.pop();
-        if(operator == ConjunctionSymbol.AND) {
-            stack.push(new AndTree(left, right));
-        }
-        else {
-            stack.push(new OrTree(left, right));
-        }
     }
 
     /**
